# openapi.yaml
openapi: 3.0.3
info:
  title: ZTN_SIM IAMaaS API
  version: "1.0.0"
  description: |
    Versioned REST contract for IAMaaS tenants and core platform operations. The schema documents
    authentication, token issuance, trust-score inspection, policy management, audit/log retrieval,
    and supporting wallet/transaction flows. It is intended to be consumed by the auth-gateway,
    policy decision point, external HMS integrations, and telecom or mobile-money pilots.
servers:
  - url: https://localhost.localdomain.com
    description: Local HTTPS (Podman + nginx)
  - url: http://localhost.localdomain.com
    description: Local HTTP (if configured)
security:
  - BearerAuth: []
    ApiKeyAuth: []
paths:
...
  /api/v1/auth/webauthn/register-begin:
    post:
      tags: [IAM]
      summary: Begin WebAuthn registration flow
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: PublicKeyCredentialCreationOptions
          content:
            application/json:
              schema:
                type: object
                description: WebAuthn registration challenge
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/auth/webauthn/register-complete:
    post:
      tags: [IAM]
      summary: Finalize WebAuthn registration and persist credentials
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: WebAuthn attestation response
      responses:
        '200': { $ref: '#/components/responses/MessageResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/auth/webauthn/assertion-begin:
    post:
      tags: [IAM]
      summary: Begin a WebAuthn assertion during login
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebauthnAssertionBeginRequest'
      responses:
        '200':
          description: PublicKeyCredentialRequestOptions
          content:
            application/json:
              schema:
                type: object
                description: WebAuthn assertion challenge
  /api/v1/auth/webauthn/assertion-complete:
    post:
      tags: [IAM]
      summary: Complete WebAuthn assertion for login
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: WebAuthn assertion response
      responses:
        '200':
          description: Assertion accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/webauthn/reset-assertion-begin:
    post:
      tags: [IAM]
      summary: Start WebAuthn recovery during password reset
      security:
        - ApiKeyAuth: []
      responses:
        '200': { $ref: '#/components/responses/MessageResponse' }
  /api/v1/auth/webauthn/reset-assertion-complete:
    post:
      tags: [IAM]
      summary: Finish WebAuthn recovery for password reset
      security:
        - ApiKeyAuth: []
      responses:
        '200': { $ref: '#/components/responses/MessageResponse' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    MessageResponse:
      description: Operation succeeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    Message:
      type: object
      properties:
        message:
          type: string
...
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        mfa_required:
          type: boolean
        trust_score:
          type: number
          format: float
...
    WebauthnAssertionBeginRequest:
      type: object
      properties:
        identifier:
          type: string
...